Test with nominal accuracy against GNU MPFR reference
Rewrite the tests to use the GNU MPFR arbitrary precision functions to compute
the reference values, and the accuracy of each back-end for the comparisons.

GNU MPFR is used through the mpfr::real template-based C++ wrapper.

The accuracy in ULPs of the implementation of the mathematical functions for
each back-end is documented in the README.md, with links to the documentation
of each library.

Clean up and reorganise the compilation flags.