Simplify the unit tests (#48)
Rewrite the test functions to simplify the code, and avoid copying the input
dataset multiple times.

Note: while the xtd library should work correctly on any version of CUDA and
ROCm that supports c++20 and concepts, the unit tests require CUDA 12.9 (to
compile correctly) and ROCm 7.1 (to support Radeon Pro).